+++ source/Application/Controller/Admin/ArticleAccessoriesAjax.php
@@ -6,9 +6,11 @@
 
 namespace OxidEsales\EshopCommunity\Application\Controller\Admin;
 
-use oxRegistry;
-use oxDb;
-use oxField;
+use OxidEsales\Eshop\Core\DatabaseProvider;
+use OxidEsales\Eshop\Core\Exception\DatabaseConnectionException;
+use OxidEsales\Eshop\Core\Field;
+use OxidEsales\Eshop\Core\Model\ListModel;
+use OxidEsales\Eshop\Core\Registry;
 
 /**
  * Class controls article assignment to accessories
@@ -23,80 +25,105 @@ class ArticleAccessoriesAjax extends \OxidEsales\Eshop\Application\Controller\Ad
     protected $_blAllowExtColumns = true;
 
     /**
+     * Container ID
+     *
+     * @var string
+     */
+    private $containerId;
+
+    /**
      * Columns array
      *
      * @var array
      */
     protected $_aColumns = ['container1' => [ // field , table,         visible, multilanguage, ident
-        ['oxartnum', 'oxarticles', 1, 0, 0],
-        ['oxtitle', 'oxarticles', 1, 1, 0],
-        ['oxean', 'oxarticles', 1, 0, 0],
-        ['oxmpn', 'oxarticles', 0, 0, 0],
-        ['oxprice', 'oxarticles', 0, 0, 0],
-        ['oxstock', 'oxarticles', 0, 0, 0],
-        ['oxid', 'oxarticles', 0, 0, 1]
+                                              ['oxartnum', 'oxarticles', 1, 0, 0],
+                                              ['oxtitle', 'oxarticles', 1, 1, 0],
+                                              ['oxean', 'oxarticles', 1, 0, 0],
+                                              ['oxmpn', 'oxarticles', 0, 0, 0],
+                                              ['oxprice', 'oxarticles', 0, 0, 0],
+                                              ['oxstock', 'oxarticles', 0, 0, 0],
+                                              ['oxid', 'oxarticles', 0, 0, 1]
     ],
-                                 'container2' => [
-                                     ['oxartnum', 'oxarticles', 1, 0, 0],
-                                     ['oxtitle', 'oxarticles', 1, 1, 0],
-                                     ['oxean', 'oxarticles', 1, 0, 0],
-                                     ['oxmpn', 'oxarticles', 0, 0, 0],
-                                     ['oxprice', 'oxarticles', 0, 0, 0],
-                                     ['oxstock', 'oxarticles', 0, 0, 0],
-                                     ['oxid', 'oxaccessoire2article', 0, 0, 1]
-                                 ]
+                            'container2' => [
+                                ['oxartnum', 'oxarticles', 1, 0, 0],
+                                ['oxtitle', 'oxarticles', 1, 1, 0],
+                                ['oxsort', 'oxaccessoire2article', 1, 1, 0],
+                                ['oxean', 'oxarticles', 1, 0, 0],
+                                ['oxmpn', 'oxarticles', 0, 0, 0],
+                                ['oxprice', 'oxarticles', 0, 0, 0],
+                                ['oxstock', 'oxarticles', 0, 0, 0],
+                                ['oxid', 'oxaccessoire2article', 0, 0, 1]
+                            ]
     ];
 
     /**
      * Returns SQL query for data to fetc
      *
      * @return string
+     * @throws DatabaseConnectionException
      */
     protected function _getQuery()
     {
-        $myConfig = $this->getConfig();
-        $sSelId = \OxidEsales\Eshop\Core\Registry::getConfig()->getRequestParameter('oxid');
-        $sSynchSelId = \OxidEsales\Eshop\Core\Registry::getConfig()->getRequestParameter('synchoxid');
-        $oDb = \OxidEsales\Eshop\Core\DatabaseProvider::getDb();
+        $myConfig = Registry::getConfig();
+        $oxidId = Registry::getConfig()->getRequestEscapedParameter('oxid');
+        $synchId = Registry::getConfig()->getRequestEscapedParameter('synchoxid');
+        $db = DatabaseProvider::getDb();
+        $this->containerId = Registry::getConfig()->getRequestEscapedParameter('cmpid');
 
-        $sArticleTable = $this->_getViewName('oxarticles');
-        $sView = $this->_getViewName('oxobject2category');
+        $articleTable = $this->_getViewName('oxarticles');
+        $object2categoryTable = $this->_getViewName('oxobject2category');
 
         // category selected or not ?
-        if (!$sSelId) {
-            $sQAdd = " from {$sArticleTable} where 1 ";
-            $sQAdd .= $myConfig->getConfigParam('blVariantsSelection') ? '' : " and {$sArticleTable}.oxparentid = '' ";
+        if (!$oxidId) {
+            $outputQuery = " from {$articleTable} where 1 ";
+            $outputQuery .= $myConfig->getConfigParam('blVariantsSelection') ? '' : " and {$articleTable}.oxparentid = '' ";
         } else {
             // selected category ?
-            if ($sSynchSelId && $sSelId != $sSynchSelId) {
+            if ($synchId && $oxidId != $synchId) {
                 $blVariantsSelectionParameter = $myConfig->getConfigParam('blVariantsSelection');
-                $sSqlIfTrue = " ( {$sArticleTable}.oxid=$sView.oxobjectid " .
-                              "or {$sArticleTable}.oxparentid=$sView.oxobjectid )";
-                $sSqlIfFals = " {$sArticleTable}.oxid=$sView.oxobjectid ";
-                $sVariantSelectionSql = $blVariantsSelectionParameter ? $sSqlIfTrue : $sSqlIfFals;
+                $trueResponse = " ( {$articleTable}.oxid=$object2categoryTable.oxobjectid " .
+                                "or {$articleTable}.oxparentid=$object2categoryTable.oxobjectid )";
+                $failResponse = " {$articleTable}.oxid=$object2categoryTable.oxobjectid ";
+                $variantSelectionSql = $blVariantsSelectionParameter ? $trueResponse : $failResponse;
 
-                $sQAdd = " from $sView left join {$sArticleTable} on {$sVariantSelectionSql}" .
-                         " where $sView.oxcatnid = " . $oDb->quote($sSelId) . " ";
+                $outputQuery = " from $object2categoryTable left join {$articleTable} on {$variantSelectionSql}" .
+                               " where $object2categoryTable.oxcatnid = " . $db->quote($oxidId) . " ";
             } else {
-                $sQAdd = " from oxaccessoire2article left join {$sArticleTable} " .
-                         "on oxaccessoire2article.oxobjectid={$sArticleTable}.oxid " .
-                         " where oxaccessoire2article.oxarticlenid = " . $oDb->quote($sSelId) . " ";
+                $outputQuery = " from oxaccessoire2article left join {$articleTable} " .
+                               "on oxaccessoire2article.oxobjectid={$articleTable}.oxid " .
+                               " where oxaccessoire2article.oxarticlenid = " . $db->quote($oxidId) . " ";
             }
         }
 
-        if ($sSynchSelId && $sSynchSelId != $sSelId) {
+        if ($synchId && $synchId != $oxidId) {
             // performance
-            $sSubSelect = " select oxaccessoire2article.oxobjectid from oxaccessoire2article ";
-            $sSubSelect .= " where oxaccessoire2article.oxarticlenid = " . $oDb->quote($sSynchSelId) . " ";
-            $sQAdd .= " and {$sArticleTable}.oxid not in ( $sSubSelect ) ";
+            $subSelect = ' select oxaccessoire2article.oxobjectid from oxaccessoire2article ';
+            $subSelect .= " where oxaccessoire2article.oxarticlenid = " . $db->quote($synchId) . " ";
+            $outputQuery .= " and {$articleTable}.oxid not in ( $subSelect )";
         }
 
         // skipping self from list
-        $sId = ($sSynchSelId) ? $sSynchSelId : $sSelId;
-        $sQAdd .= " and {$sArticleTable}.oxid != " . $oDb->quote($sId) . " ";
+        $sId = ($synchId) ? $synchId : $oxidId;
+        $outputQuery .= " and {$articleTable}.oxid != " . $db->quote($sId) . " ";
 
         // creating AJAX component
-        return $sQAdd;
+        return $outputQuery;
+    }
+
+
+    /**
+     * overide default sorting and replace it with OXSORT field
+     *
+     * @return string
+     */
+    protected function _getSorting()
+    {
+        if ($this->containerId == 'container2') {
+            return ' order by _2,_0';
+        } else {
+            return ' order by _' . $this->_getSortCol() . ' ' . $this->_getSortDir() . ' ';
+        }
     }
 
     /**
@@ -152,5 +179,78 @@ class ArticleAccessoriesAjax extends \OxidEsales\Eshop\Application\Controller\Ad
      */
     protected function onArticleAccessoryRelationChange($article)
     {
+    }
+
+
+    /**
+     * Applies sorting for Accessories list
+     */
+    public function sortAccessoriesList()
+    {
+        $oxidRelationId = Registry::getConfig()->getRequestEscapedParameter('oxid');
+        $selectedIdForSort = Registry::getConfig()->getRequestEscapedParameter('sortoxid');
+        $sortDirection = Registry::getConfig()->getRequestEscapedParameter('direction');
+
+        $accessoriesList = oxNew(ListModel::class);
+        $accessoriesList->init("oxbase", "oxaccessoire2article");
+        $sortQuery = "select * from  oxaccessoire2article where OXARTICLENID= " . DatabaseProvider::getDb()->quote($oxidRelationId) . " order by oxsort,oxid";
+        $accessoriesList->selectString($sortQuery);
+
+
+        $rebuildList = $this->rebuildAccessoriesSortIndexes($accessoriesList);
+
+        if (($selectedPosition = array_search($selectedIdForSort, $rebuildList)) !== false) {
+            $selectedSortRecord = $accessoriesList->offsetGet($rebuildList[$selectedPosition]);
+            $currentPosition = $selectedSortRecord->oxaccessoire2article__oxsort->value;
+
+            // get current selected row sort position
+
+            if (($sortDirection == 'up' && $currentPosition > 0) || ($sortDirection == 'down' && $currentPosition < count($rebuildList) - 1)) {
+                $newPosition = ($sortDirection == 'up') ? ($currentPosition - 1) : ($currentPosition + 1);
+
+                // exchanging indexes
+                $currentRecord = $accessoriesList->offsetGet($rebuildList[$currentPosition]);
+                $newRecord = $accessoriesList->offsetGet($rebuildList[$newPosition]);
+
+                $currentRecord->oxaccessoire2article__oxsort = new Field($newPosition);
+                $newRecord->oxaccessoire2article__oxsort = new Field($currentPosition);
+                $currentRecord->save();
+                $newRecord->save();
+            }
+        }
+
+        $outputQuery = $this->_getQuery();
+
+        $normalQuery = 'select ' . $this->_getQueryCols() . $outputQuery;
+        $countQuery = 'select count( * ) ' . $outputQuery;
+
+        $this->_outputResponse($this->_getData($countQuery, $normalQuery));
+    }
+
+
+    /**
+     * rebuild Accessories sort indexes
+     *
+     * @param ListModel $inputList
+     *
+     * @return array
+     */
+    private function rebuildAccessoriesSortIndexes(ListModel $inputList): array
+    {
+
+        $counter = 0;
+        $outputList = [];
+        foreach ($inputList as $key => $value) {
+            if (isset($value->oxaccessoire2article__oxsort)) {
+                if ($value->oxaccessoire2article__oxsort->value != $counter) {
+                    $value->oxaccessoire2article__oxsort = new Field($counter);
+                    $value->save();
+                }
+            }
+            $outputList[$counter] = $key;
+            $counter++;
+        }
+
+        return $outputList;
     }
 }